src/requirements.txt
Django==3.0.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mongoengine==0.28.2
djongo==1.3.3
pymongo==3.11.2
requests==2.24.0